Linux: Tons of warning cleanup plus workaround a nasty crash problem in StatusBarMessage.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4234 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/State.cpp
@@ -315,7 +315,7 @@ void LoadStateCallback(u64 userdata, int cyclesLate)
 		buffer = new u8[sz];
 
 		int x;
-		if ((x = (int)fread(buffer, 1, sz, f)) != sz)
+		if ((x = (int)fread(buffer, 1, sz, f)) != (int)sz)
 			PanicAlert("wtf? %d %d", x, sz);
 	}
